Download, extract and open a zipped Access database from a CKAN dataset.

The extracted file will be kept in destdir. If a file with the expected filename already exists in destdir, this file will be used. To force a fresh download, remove or rename the file in destdir.

dl_mdbzip(
  resource_id,
  destdir = here::here("data"),
  dateformat = "%m-%d-%Y",
  as.is = TRUE,
  verbose = get_tsc_verbose()
)

Arguments

resource_id
The CKAN resource ID of a zipped Access DB.

destdir
The local destination directory for the extracted file, will be created if not existing. Default: here::here("data").

dateformat
The parameter dateformat for Hmisc::mdb.get(), default: %Y-%m-%d.

as.is
The parameter as.is for Hmisc::mdb.get(), default: TRUE.

verbose
Whether to show verbose messages. Default: get_tsc_verbose.

Value

The Hmisc::mdb.get connection.

See also

Examples

if (FALSE) {
ckanr::ckanr_setup(
  url = Sys.getenv("CKAN_URL"), key = Sys.getenv("CKAN_API_KEY")
)
tfa_data <- dl_mdbzip("66efb68d-8f05-4bfc-af14-5d1a381d0cf2", verbose = TRUE)
}